Fixed Add Friend button through messages
