+++ src/components/PublicMessages/MessageList.js
@@ -1,44 +1,23 @@
 //Author: Gerson Diketama
 //Purpose: Card to display all the the message from API/ GET
 
-import React, { useState } from 'react'
+import React from 'react'
 import {AiFillDelete} from "react-icons/ai"
-import {AiTwotoneEdit} from "react-icons/ai"
 import { useHistory } from 'react-router'
 import { useEffect } from 'react/cjs/react.development'
-import { getFriendsById } from '../friends/FriendManager'
-import { MessageEdit } from './PublicForm'
-import { getPublicMessages } from './publicMessageManager'
+import "../articles/ArticleCardDesign.css"
 
 
-const MessageList = ({data, handledelete, handleAddFriend}) => {
+const MessageList = ({data, handledelete, handleAddFriend, friends, getFriendsList}) => {
 
     const history = useHistory();
 
-    const [friendId, changeFriend] = useState([]);
-    const [disable, setDisable] = useState(false);
-
-    let arrayTaco = [];
-
-    const stateTaco = () => {
-        arrayTaco = [];
-
-        getFriendsById(parseInt(sessionStorage.getItem("nutshell_user"))).then(res => {
-            res.forEach(taco => {
-                arrayTaco.push(taco.userId);
-            })
-
-            changeFriend(arrayTaco);
-        }
-        )
-    }
-
     useEffect(() => {
-        stateTaco();
-    }, [disable]);
+        getFriendsList();
+    }, []);
 
 
-    if(friendId.includes(data.user.id) === true){
+    if(friends.includes(data.user.id) === true){
             return (
                 <div>
                     <h2>{data.user.name}</h2>
@@ -54,7 +33,6 @@ const MessageList = ({data, handledelete, handleAddFriend}) => {
                 <h4>{data.post}</h4>
                 <button onClick={(event) => {
                     handleAddFriend(data.user.id, event)
-                    setDisable(true);
                     }}>Add as friend</button>
             </div>
         )
