+++ src/components/PublicMessages/PublicMessages.js
@@ -4,17 +4,17 @@
 import React, {useEffect } from "react";
 import { deleteMessages } from "./publicMessageManager";
 import MessageList from "./MessageList";
-import SentMessages from "./SentMessages";
 import { addFriends, getFriendsById } from "../friends/FriendManager";
 import { getUserById } from "../users/UserManager";
 
 
-export const PublicMessages = ({messages, getPublicMessages}) => {
+export const PublicMessages = ({messages, getPublicMessages, friends, getFriendsList}) => {
 
 // THIS USE EFFECT WATCHES FOR ANY CHANGES MADE ON APPLICATION VIEWS AND RE-RENDER
 
   useEffect(() => {
     getPublicMessages();
+    getFriendsList();
   }, []);
 
   const handleDelete = (messageId) =>
@@ -24,7 +24,7 @@ export const PublicMessages = ({messages, getPublicMessages}) => {
 
   }
 
-  const handleAddFriend = (input, button) => {
+  const handleAddFriend = (input) => {
 
     getUserById(parseInt(sessionStorage.getItem("nutshell_user"))).then(res => {
       console.log(res);
@@ -34,13 +34,13 @@ export const PublicMessages = ({messages, getPublicMessages}) => {
         currentUserId: res.id,
         currentUserName: res.name
       }
-      addFriends(friendObj).then(() => getFriendsById(parseInt(sessionStorage.getItem("nutshell_user"))))
+      addFriends(friendObj).then(() => getFriendsList());
     })
   }
   return (
     <>
       {messages.map((allMessages) => (
-        <MessageList key={allMessages.id} data={allMessages} handledelete={handleDelete} handleAddFriend={handleAddFriend}/>
+        <MessageList key={allMessages.id} data={allMessages} handledelete={handleDelete} handleAddFriend={handleAddFriend} friends={friends} getFriendsList={getFriendsList}/>
       ))}
       
     </>
