+++ src/components/PublicMessages/SentMessages.js
@@ -2,7 +2,7 @@
 
 //Purpose: This is a public message component, allows users to send messages back and forwards, and I am using React-icons for icons. 
 
-import React, { useState, useEffect } from "react";
+import React, { useState} from "react";
 import { addPublicMessages } from "./publicMessageManager";
 import {AiOutlineSend} from "react-icons/ai"
 import { getUserById } from "../users/UserManager";
